extracted default_behavior callables into their specific files
